+++ teamcity/datadog_checks/teamcity/common.py
@@ -1,53 +1,88 @@
 # (C) Datadog, Inc. 2022-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import re
 import time
 from copy import deepcopy
 
 import requests
 
-from datadog_checks.base import AgentCheck
+from datadog_checks.base import ConfigurationError
 
-EVENT_STATUS_MAP = {"SUCCESS": "successful", "FAILURE": "failed"}
+from .constants import BUILD_EVENT, DEPLOYMENT_EVENT, EVENT_STATUS_MAP, RESOURCE_URL_MAP
 
-SERVICE_CHECK_STATUS_MAP = {"SUCCESS": AgentCheck.OK, "FAILURE": AgentCheck.CRITICAL}
 
-RESOURCE_URL_MAP = {
-    "build_configs": "{base_url}/app/rest/buildTypes?locator=project(id:{project_id})",
-    "build_config": "{base_url}/app/rest/buildTypes/id:{build_conf}",
-    "new_builds": "{base_url}/app/rest/builds/?locator=buildType:{build_conf},sinceBuild:id:{since_build},"
-    "state:finished,defaultFilter:false",
-    "started_builds": "{base_url}/app/rest/builds/?locator=buildType:{build_conf},sinceBuild:id:{since_build},"
-    "state:finished,defaultFilter:false",
-    "last_build": "{base_url}/app/rest/builds/?locator=buildType:{build_conf},count:1",
-    "build_stats": "{base_url}/app/rest/builds/buildType:{build_conf},buildId:{build_id}/statistics",
-    "agent_usage_stats": "{base_url}/app/rest/builds?locator=agent:(id:{agent_id})&fields=build(id:{build_id},"
-    "buildTypeId:{build_conf})",
-    "test_occurrences": "{base_url}/app/rest/testOccurrences?locator=build:{build_id}",
-    "build_problems": "{base_url}/app/rest/problemOccurrences?locator=build:(id:{build_id})",
-}
+def construct_build_configs_filter(check):
+    """
+    Construct build configuration filter
+    """
+    excluded_build_configs = set()
+    included_build_configs = set()
+    for project in check.monitored_build_configs:
+        config = check.monitored_build_configs.get(project)
+        # collect all build configs in project
+        if isinstance(config, dict):
+            exclude_list = config.get('exclude', [])
+            include_list = config.get('include', [])
+            if exclude_list:
+                excluded_build_configs.update(exclude_list)
+            if include_list:
+                for include_bc in include_list:
+                    if include_bc not in excluded_build_configs:
+                        included_build_configs.update([include_bc])
+        elif config is None:
+            continue
+        else:
+            raise ConfigurationError(
+                "`project` must be either an empty mapping to collect all build configurations in the project"
+                "or a mapping of keys `include` and/or `exclude` lists of build configurations."
+            )
+    return excluded_build_configs, included_build_configs
 
-EVENT_BASE = {"timestamp": None, "source_type_name": "teamcity", "host": None, "tags": []}
 
-DEPLOYMENT_EVENT = {
-    "event_type": "teamcity_deployment",
-    "msg_title": "{instance_name} deployed to {host} {build_status}",
-    "msg_text": "Build Number: {build_number}\n\nMore Info: {build_webUrl}",
-    "tags": ["teamcity:deployment"],
-}
+def should_include_build_config(check, build_config):
+    """
+    Return `True` if the build_config is included, otherwise `False`
+    """
+    exclude_filter, include_filter = construct_build_configs_filter(check)
+    include_match = False
+    exclude_match = False
+    # If no filters configured, include everything
+    if not exclude_filter and not include_filter:
+        return True
+    if exclude_filter:
+        for pattern in exclude_filter:
+            if re.search(re.compile(pattern), build_config):
+                exclude_match = True
+    if include_filter:
+        for pattern in include_filter:
+            if re.search(re.compile(pattern), build_config):
+                include_match = True
 
-BUILD_EVENT = {
-    "event_type": "build",
-    "msg_title": "Build for {instance_name} {build_status}",
-    "msg_text": "Build Number: {build_number}\nDeployed To: {host}\n\nMore Info: {build_webUrl}",
-    "tags": ["build"],
-}
-
-DEPLOYMENT_EVENT.update(EVENT_BASE)
-BUILD_EVENT.update(EVENT_BASE)
+    # Include everything except in excluded_bc
+    if exclude_filter and not include_filter:
+        return not exclude_match
+    # Include only what's defined in included_bc
+    if include_filter and not exclude_filter:
+        return include_match
+    # If both include and exclude filters are configured
+    if include_filter and exclude_filter:
+        # If filter overlap or in neither filter, exclude
+        if (include_match and exclude_match) or (not include_match and not exclude_match):
+            return False
+        # Only matches include filter, include
+        if include_match and not exclude_match:
+            return include_match
+        # Only matches exclude filter, exclude
+        if exclude_match and not include_match:
+            return not exclude_match
+    return True
 
 
 def construct_event(is_deployment, instance_name, host, new_build, build_tags):
+    """
+    Construct event based on build status.
+    """
     build_id = new_build['id']
     build_number = new_build['number']
     build_config = new_build['buildTypeId']
@@ -97,40 +132,3 @@ def get_response(check, resource, **kwargs):
     except Exception:
         check.log.exception("Couldn't fetch resource %s, unhandled exception", resource_name)
         raise
-
-
-class Build(object):
-    def __init__(self, build_id):
-        self.build_id = build_id
-
-
-class BuildConfig(object):
-    def __init__(self, build_config_id):
-        self.build_config_id = build_config_id
-        self.project_id = None
-        self.last_build_id = None
-
-
-class BuildConfigs(BuildConfig):
-    def __init__(self):
-        self.build_configs = {}
-
-    def get_build_configs(self):
-        return deepcopy(self.build_configs)
-
-    def set_build_config(self, build_type_id, project_id):
-        if not self.build_configs.get(build_type_id):
-            self.build_configs[build_type_id] = BuildConfig(build_type_id)
-            self.build_configs[build_type_id].project_id = project_id
-
-    def get_build_config(self, build_type_id):
-        return self.build_configs.get(build_type_id, None)
-
-    def set_last_build_id(self, build_type_id, build_id):
-        self.build_configs[build_type_id].last_build_id = build_id
-
-    def get_last_build_id(self, build_type_id):
-        build_config = self.build_configs.get(build_type_id, None)
-        if build_config:
-            return build_config.last_build_id
-        return None
